fix SSH host keys for all formats
Root cause of the breakage is that I need to call ssh-keygen with the
legacy PEM format (-m option), which is what ParsePrivateKey expects.
This is visible as "BEGIN DSA PRIVATE KEY" instead of "BEGIN OPENSSH
PRIVATE KEY" in the PEM files generated.

While I'm at it, clean up LoadHostKeys and make all errors there
non-fatal as long as at least one key is successfully loaded.