fix username handling
user.Current() doesn't work when cross-compiling (cgo disabled) because
it can't call getpwnam in glibc, and apparently start-stop-daemon
doesn't set enough environment variables for the fallback
implementations to work.

Add a command line option for the SSH user and only try user.Current()
or $USER if that's unset. Move this logic to main instead of server.
Print the expected user when the server starts to listen.

Make NewServer return a pointer like it should so that the auth callback
actually points to the correct object.