replace magic_packet.go with import "github.com/sabhiram/go-wol"
Importing ini already added a bunch of clutter to go.mod/go.sum so
I might as well be consistent and do the same thing with magic packet
stuff.

"go mod" is smart enough to only vendor and compile the files I need, so
the "useless" requirements don't actually pull in extra code.